# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Closechain AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/subcontractors/{subcontractorId}/documents:
    get:
      operationId: listDocumentSlots
      tags: [documents]
      summary: List document slots for a subcontractor
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: subcontractorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentSlot"
    post:
      operationId: addDocumentSlot
      tags: [documents]
      summary: Add a custom document requirement
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: subcontractorId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddDocumentSlotBody"
      responses:
        "201":
          description: Document slot added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentSlot"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /documents/{documentSlotId}:
    patch:
      operationId: updateDocumentSlot
      tags: [documents]
      summary: Update a document slot (status, file)
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: documentSlotId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentSlotBody"
      responses:
        "200":
          description: Updated document slot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentSlot"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteDocumentSlot
      tags: [documents]
      summary: Remove a document requirement
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: documentSlotId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /projects/{projectId}/documents:
    get:
      operationId: listAllProjectDocuments
      tags: [documents]
      summary: List all document slots across all subcontractors for a project
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [not_submitted, uploaded, approved]
        - name: documentType
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: All document slots for the project
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentSlotWithSubcontractor"
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — Bearer <sid>.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    DocumentSlot:
      type: object
      required: [id, subcontractorId, documentType, status, createdAt]
      properties:
        id:
          type: integer
        subcontractorId:
          type: integer
        documentType:
          type: string
        packageSection:
          type: [string, 'null']
        status:
          type: string
          enum: [not_submitted, uploaded, approved]
        filePath:
          type: [string, 'null']
        fileName:
          type: [string, 'null']
        uploadedAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
    DocumentSlotWithSubcontractor:
      type: object
      required: [id, subcontractorId, documentType, status, vendorName, csiCode, createdAt]
      properties:
        id:
          type: integer
        subcontractorId:
          type: integer
        documentType:
          type: string
        packageSection:
          type: [string, 'null']
        status:
          type: string
          enum: [not_submitted, uploaded, approved]
        filePath:
          type: [string, 'null']
        fileName:
          type: [string, 'null']
        uploadedAt:
          type: [string, 'null']
          format: date-time
        vendorName:
          type: string
        csiCode:
          type: string
        createdAt:
          type: string
          format: date-time
    AddDocumentSlotBody:
      type: object
      required: [documentType]
      properties:
        documentType:
          type: string
          minLength: 1
    UpdateDocumentSlotBody:
      type: object
      properties:
        status:
          type: string
          enum: [not_submitted, uploaded, approved]
        filePath:
          type: string
        fileName:
          type: string
...