# lib/api-zod/src/generated/types/documentSlot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Closechain AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentSlotStatus } from "./documentSlotStatus";

export interface DocumentSlot {
  id: number;
  subcontractorId: number;
  documentType: string;
  /** @nullable */
  packageSection?: string | null;
  status: DocumentSlotStatus;
  /** @nullable */
  filePath?: string | null;
  /** @nullable */
  fileName?: string | null;
  /** @nullable */
  uploadedAt?: Date | null;
  createdAt: Date;
}