# lib/api-zod/src/generated/types/documentSlotWithSubcontractor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Closechain AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentSlotWithSubcontractorStatus } from "./documentSlotWithSubcontractorStatus";

export interface DocumentSlotWithSubcontractor {
  id: number;
  subcontractorId: number;
  documentType: string;
  /** @nullable */
  packageSection?: string | null;
  status: DocumentSlotWithSubcontractorStatus;
  /** @nullable */
  filePath?: string | null;
  /** @nullable */
  fileName?: string | null;
  /** @nullable */
  uploadedAt?: Date | null;
  vendorName: string;
  csiCode: string;
  createdAt: Date;
}